#22-24 ajax delete

+++ dotinstall/cakephp/blog/app/Controller/PostsController.php
@@ -44,9 +44,16 @@ class PostsController extends AppController {
         if ($this->request->is('get')) {
             throw new MethodNotAllowedException();
         }
-        if ($this->Post->delete($id)) {
-            $this->Session->setFlash('Deleted!');
-            $this->redirect(array('action' => 'index'));
+        if ($this->request->is('ajax')) {
+            if ($this->Post->delete($id)) {
+                $this->autoRender = false;
+                $this->autoLayout = false;
+                $response = array('id' => $id);
+                $this->header('Content-Type: application/json');
+                echo json_encode($response);
+                exit();
+            }
         }
+        $this->redirect(array('action' => 'index'));
     }
 }
